+++ src/simulation/elements/FRZW.cpp
@@ -60,7 +60,7 @@ int Element_FRZW::update(UPDATE_FUNC_ARGS)
 					sim->part_change_type(ID(r),x+rx,y+ry,PT_FRZW);
 				}
 			}
-	if ((parts[i].life==0 && !(random_gen()%192)) || (100-(parts[i].life))>random_gen()%50000 )
+	if ((parts[i].life==0 && !(random_gen()%192)) || (100-(parts[i].life)) > (int)(random_gen()%50000))
 	{
 		sim->part_change_type(i,x,y,PT_ICEI);
 		parts[i].ctype=PT_FRZW;
