+++ src/simulation/elements/ISZS.cpp
@@ -48,7 +48,7 @@ Element_ISZS::Element_ISZS()
 int Element_ISZS::update(UPDATE_FUNC_ARGS)
  { // for both ISZS and ISOZ
 	float rr, rrr;
-	if (!(random_gen()%200) && ((int)(-4.0f*(sim->pv[y/CELL][x/CELL])))>(random_gen()%1000))
+	if (!(random_gen()%200) && (-4.0f*(sim->pv[y/CELL][x/CELL])) > (random_gen()%1000))
 	{
 		sim->create_part(i, x, y, PT_PHOT);
 		rr = (random_gen()%228+128)/127.0f;
