+++ src/simulation/elements/LIGH.cpp
@@ -91,7 +91,7 @@ int Element_LIGH::update(UPDATE_FUNC_ARGS)
 				rt = TYP(r);
 				if ((surround_space || sim->elements[rt].Explosive) &&
 				    (rt!=PT_SPNG || parts[ID(r)].life==0) &&
-					sim->elements[rt].Flammable && (sim->elements[rt].Flammable + (int)(sim->pv[(y+ry)/CELL][(x+rx)/CELL]*10.0f))>(random_gen()%1000))
+					sim->elements[rt].Flammable && (sim->elements[rt].Flammable + (sim->pv[(y+ry)/CELL][(x+rx)/CELL]*10.0f)) > (random_gen()%1000))
 				{
 					sim->part_change_type(ID(r),x+rx,y+ry,PT_FIRE);
 					parts[ID(r)].temp = restrict_flt(sim->elements[PT_FIRE].Temperature + (sim->elements[rt].Flammable/2), MIN_TEMP, MAX_TEMP);
@@ -278,7 +278,7 @@ bool Element_LIGH::create_LIGH(Simulation * sim, int x, int y, int c, int temp, 
 		sim->parts[p].tmp = tmp;
 		if (last)
 		{
-			sim->parts[p].tmp2=1+(random_gen()%200>tmp2*tmp2/10+60);
+			sim->parts[p].tmp2=1+((int)(random_gen()%200) > tmp2*tmp2/10+60);
 			sim->parts[p].life=(int)(life/1.5-random_gen()%2);
 		}
 		else
